Added attempt to integrate CCP, no longer build error, but calls don't come in
